Try to find bug

+++ ft_printf.h
@@ -16,6 +16,7 @@
 # include <unistd.h>
 # include <stdlib.h>
 # include <stdarg.h>
+# include <stdint.h>
 //# include "libft.h"
 
 # define FLAG_LIMIT 5
@@ -63,9 +64,9 @@ void print_hex_X(unsigned int nbr);
 void print_hex_o(unsigned int nbr);
 void print_hex_O(unsigned int nbr);
 void check_stars(t_argc *params, va_list ap);
-void print_int_depend_length(intmax_t d, char *length, t_argc params);
-void print_int_params_left(intmax_t d, t_argc params, int zeros, int spaces);
-void print_int_params_right(intmax_t d, t_argc params, int zeros, int spaces);
+void print_int_depend_length(long long d, char *length, t_argc params);
+void print_int_params_left(long long d, t_argc params, int zeros, int spaces);
+void print_int_params_right(long long d, t_argc params, int zeros, int spaces);
 
 
 int		ft_strcmp(const char *s1, const char *s2);
